# docs/technical/api-documentation.yaml
openapi: 3.0.3
info:
  title: Post-Facto Coding Review API
  description: |
    HIPAA-compliant API for automated medical coding review and revenue optimization.

    ## Overview
    This API allows healthcare providers to submit clinical notes and billed codes for automated review,
    identifying missed billing opportunities using AI-powered analysis.

    ## Authentication
    All API endpoints require JWT authentication via Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## HIPAA Compliance
    All PHI is processed through Amazon Comprehend Medical for de-identification before AI analysis.
    Data is encrypted at rest and in transit (TLS 1.3).

  version: 1.0.0
  contact:
    name: RevRX Support
    email: support@revrx.com
  license:
    name: Proprietary
servers:
  - url: https://api.revrx.com/v1
    description: Production server
  - url: https://staging-api.revrx.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email verification
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - fullName
              properties:
                email:
                  type: string
                  format: email
                  example: doctor@clinic.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                fullName:
                  type: string
                  example: Dr. Jane Smith
                organization:
                  type: string
                  example: City Medical Center
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: JWT access token
                  refreshToken:
                    type: string
                    description: JWT refresh token
                  expiresIn:
                    type: integer
                    description: Token expiration in seconds
                    example: 3600
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        organization:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
        subscriptionStatus:
          type: string
          enum: [trial, active, cancelled, expired]
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []